Try moving code to mixin
